Factor out utils relating to base64 encoding playgrounds
